+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [Optimize: support delete config for ConfigurationProperties bean  ](https://github.com/Tencent/spring-cloud-tencent/pull/756)
 - [Feature: support nacos registry](https://github.com/Tencent/spring-cloud-tencent/pull/758)
 - [Test:add sct-stater-polaris-router junit.](https://github.com/Tencent/spring-cloud-tencent/pull/768)
+- [Fix issue: prompt no registration if web dependence was not imported.](https://github.com/Tencent/spring-cloud-tencent/pull/778)